abstract collector/exporter; float32

+++ src/collectors/zapi/plugins/flexgroup/flexgroup.go
@@ -93,7 +93,7 @@ func (p *FlexGroup) Run(data *matrix.Matrix) ([]*matrix.Matrix, error) {
 		for _, metric := range n.GetMetrics() {
 			if strings.Contains(metric.Display, "percent") {
 				if value, has := n.GetValue(metric, instance); has {
-					n.SetValue(metric, instance, value / float64(count))
+					n.SetValue(metric, instance, value / float32(count))
 				}
 			}
 		}
